Perl mod pkgbuilds use $_ddir now.
With pacman 4.1 $srcdir is no longer set outside of PKGBUILD
functions. Yipee! This breaks all of my PKGBUILDs. Now we
switch to storing the distribution's tarballed directory in the
_ddir custom variable. We cd into $srcdir/$_ddir at the
beginning of every func.

Custom variable mechanics have changed in bin/pbfields as
well. Instead of hard-coding _dir as an accepted field, any
PKGDATA entry with a name starting with underscore (_) is
considered a custom variable and printed in the PKGBUILD.